app: Split modules from inbound and outbound
The app::inbound and app::outbound modules have been growing in size
over time as features are added.

This change splits these large many-module files up so that the inbound
and outbound stack specifications can also be moved into these modules.
