+++ src/app/inbound/mod.rs
@@ -1,0 +1,9 @@
+mod endpoint;
+pub(super) mod orig_proto_downgrade;
+pub(super) mod rewrite_loopback_addr;
+#[allow(dead_code)] // TODO #2597
+mod set_client_id_on_req;
+#[allow(dead_code)] // TODO #2597
+mod set_remote_ip_on_req;
+
+pub use self::endpoint::{Endpoint, RecognizeEndpoint};
