+++ src/app/outbound/mod.rs
@@ -1,0 +1,11 @@
+#[allow(dead_code)] // TODO #2597
+mod add_remote_ip_on_rsp;
+#[allow(dead_code)] // TODO #2597
+mod add_server_id_on_rsp;
+pub(super) mod discovery;
+pub(super) mod endpoint;
+pub(super) mod orig_proto_upgrade;
+pub(super) mod require_identity_on_endpoint;
+
+pub(super) use self::endpoint::Endpoint;
+pub(super) use self::require_identity_on_endpoint::RequireIdentityError;
